% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.metadata-package.R
\docType{package}
\name{dc.metadata-package}
\alias{dc.metadata}
\alias{dc.metadata-package}
\title{dc.metadata: Gets Metadata from SDAD Data Commons}
\description{
Allows user to easily get metadata from the SDAD Data Commons.  Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Joanna Schroeder \email{js2mr@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
